(feat) - start migrating hooks to core
