+++ src/index.js
@@ -6,3 +6,4 @@ export { createContext } from './create-context';
 export { toChildArray } from './diff/children';
 export { unmount as _unmount } from './diff';
 export { default as options } from './options';
+export * from './hooks';
